requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
requests>=2.28.0
psutil>=5.9.0
tensorboard>=2.13.0
